(QENG-1453) Make use of Beaker's epel_url option
Previously we were depending on the "use-el-extras" feature of Beaker
but we were not taking full advantage of its ability to point at an EPEL
mirror.  We have sporadically seen failures based on the availability of
the mirrors at kernel.org.

This commit infers if we are running within the internal network by
looking for a well known environment variable and, if so, adds the
appropriate beaker configuration to use our internal epel mirror.