(PUP-8659) Bypass MultiJSON for catalog serialization
In order to properly downgrade to PSON for catalogs with binary data we
expect the JSON serializer to raise if it contains content in the wrong
encoding. JrJackson will not do this and instead will replace invalid
UTF-8 with the Unicode replacement character.

This commit moves us back to using the built in JSON serializer to
retain the raise-on-invalid-sequence behavior.