(PUP-8656) First pass of respecting CRL chains
Previously, we could not load CRL paths that contained multiple CRLs
causing users who wished to use intermediate CAs to have to disable
CRL checking.

This assumes that the CRL path contains a file with one or more PEM
encoded CRLs concantenated together. It will attempt to parse and add
each in turn to the SSL Store used by the Host when validating HTTPS
connections.

This also moves us from mocking out the SSL Store and asserting we call
specific methods on it to creating an actual Store and chained CA
bundles with revoked and unrevoked nodes at each level in our tests. We
can then assert that the Store verifies certs accordingly.

To enable this there is an additional test utility PuppetSpec::SSL that
will help create keys, certs, and crls in memory for testing.