(PUP-8656) Failing tests for SSL::Host usage of CRL chains
This moves us from mocking out the SSL Store and asserting we call
specific methods on it to creating an actual Store and chained CA
bundles with revoked and unrevoked nodes at each level. We can then
assert that the Store verifies certs accordingly.

To enable this there is an additional test utility PuppetSpec::SSL that
will help create keys, certs, and crls in memory for testing.