(PUP-8656) Move side effects out of SSL Store building
Previously, the `SSL::Host#build_ssl_store` method would make several
settings lookups (sometimes for the same setting) that could change
during execution, potentially make a network request to download the
CRL, and handled the case if the CRL was asked for, but could not be
downloaded.

This moves setting lookups to happen once during object initialization
and updates specs as necessary.

This moves the potential downloading of a CRL chain to a separate
method that also handles if the CRL cannot be retrieved.

This moves loading the CRL chain to a separate method that handles if
the chain is malformed.