Just tasks
This adds two new Lexers, CatalogLexer and TaskLexer that subclass
Lexer2 and updates the Parsers to use Puppet.lookup(:current_lexer)
instead of hard coding references to a single lexer.

It pupts CatalogLexer in the bootstrap context by default (which
mirrors Puppet's default setting of `tasks: false`.

It has to be in the bootstrap context because some PCore is parsed
on file `require`.

This should maybe be a larger split between a catalog and a tasks
compiler, with splits all the way down to the lexer....

Because the Evaluator, Loaders, and a handful of functions check to see
if tasks are enabled we also populate a context value of `tasks` and
update all of the above to check the context.

This still allows setting the tasks setting but now gives a deprecation
warning and sets the value only for the current thread.